+++ app/src/main/java/com/example/mayingnan/project301/Task.java
@@ -16,7 +16,7 @@ public class Task {
     private String taskProvider;
     private String taskStatus;
     private String taskAddress;
-    private Float lowestBid;
+    private Double lowestBid;
 
     private String taskID;
 
@@ -48,7 +48,7 @@ public class Task {
         this.taskPhoto=null;
         this.taskIdealPrice=null;
         this.taskDateTime=null;
-        this.lowestBid = (float)0;
+        this.lowestBid = 0.0;
     }
 
     public Task(String taskName, String taskDetails, String taskRequester, String taskProvider,
@@ -63,7 +63,7 @@ public class Task {
         this.taskPhoto=taskPhoto;
         this.taskIdealPrice=null;
         this.taskDateTime=null;
-        this.lowestBid = (float)0.0;
+        this.lowestBid = 0.0;
     }
 
     public void requesterAcceptsBid(){
@@ -124,14 +124,14 @@ public class Task {
     public ArrayList<Bid> getTaskBidList(){
         return taskBidList;
     }
-    public float getLowestBid() {return this.lowestBid;}
+    public Double getLowestBid() {return this.lowestBid;}
     public Photo getTaskPhoto(){
         return taskPhoto;
     }
 
     public boolean addBid(Bid bid){
         if (bid.getBidAmount() < this.lowestBid){
-            this.lowestBid = bid.getBidAmount();
+            this.lowestBid = (Double) bid.getBidAmount();
         }
         for (int i = 0; i<taskBidList.size();i++){
             String selectedItemProvider = taskBidList.get(i).getProviderId();
